py3: don't use ensure_ascii kwarg of JSONCodec
ensure_ascii kwarg is not documented neither used in Taurus.
Seems like this kwarg when used JSONCodec.decode additionally encodes to
bytes using utf-8. This breaks the usage in Sardana. The normal use of
JSONCodec seems to be enough with encoding and decoding (it already encodes
and decodes with utf-8). Remove usage of ensure_ascii.